Exercise 1
Exercise 1: Sign Extension
The purpose of this exercise is to get hands-on with C’s bitwise
operators. Baseline
source code, which includes a test harness (in main()), is provided.
Begin by setting
the active project to “Exercise1”. Then implement the body of function
signextend(),
in template file signextend.c, so that it:
• Examines the 12-bit “2’s complement” input value in the uint16_t
parameter.3
• Returns the correct 16-bit “2’s complement” value for all possible
input values.